chore: migrate suggestions to use TS
